chore: refactor dialogs and favorites
