+++ src/BSH.MainApp/Models/BrowserContentMode.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Alexander Seeliger. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0.
+
+namespace BSH.MainApp.Models;
+
+public enum BrowserContentMode
+{
+    Folder,
+    Search
+}
