+++ src/BSH.MainApp/Services/BrowserDialogService.cs
@@ -1,0 +1,167 @@
+// Copyright (c) Alexander Seeliger. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0.
+
+using Brightbits.BSH.Engine.Models;
+using BSH.MainApp.Contracts.Services;
+using BSH.MainApp.Models;
+using CommunityToolkit.WinUI;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Windows.UI.Popups;
+
+namespace BSH.MainApp.Services;
+
+public class BrowserDialogService : IBrowserDialogService
+{
+    private readonly IPresentationService presentationService;
+
+    public BrowserDialogService(IPresentationService presentationService)
+    {
+        this.presentationService = presentationService;
+    }
+
+    public async Task<IReadOnlyList<string>> ShowDeleteBackupsWindowAsync(IReadOnlyList<VersionDetails> versions)
+    {
+        return await App.MainWindow.DispatcherQueue.EnqueueAsync(async () =>
+        {
+            var versionList = new ListView
+            {
+                SelectionMode = ListViewSelectionMode.Multiple,
+                ItemsSource = versions,
+                DisplayMemberPath = nameof(VersionDetails.CreationDate),
+                MaxHeight = 360
+            };
+
+            var dialog = new ContentDialog
+            {
+                XamlRoot = App.MainWindow.Content.XamlRoot,
+                Title = "Delete backups",
+                Content = versionList,
+                PrimaryButtonText = "Delete",
+                CloseButtonText = "Cancel",
+                DefaultButton = ContentDialogButton.Close
+            };
+
+            var result = await dialog.ShowAsync();
+            if (result != ContentDialogResult.Primary)
+            {
+                return Array.Empty<string>();
+            }
+
+            var selected = versionList.SelectedItems.Cast<VersionDetails>().Select(x => x.Id).ToArray();
+            if (selected.Length == 0)
+            {
+                return Array.Empty<string>();
+            }
+
+            var confirmDialog = new ContentDialog
+            {
+                XamlRoot = App.MainWindow.Content.XamlRoot,
+                Title = "Delete backups",
+                Content = new TextBlock { Text = "Are you sure you want to delete the selected backups?" },
+                PrimaryButtonText = "Yes",
+                CloseButtonText = "No"
+            };
+
+            return await confirmDialog.ShowAsync() == ContentDialogResult.Primary ? selected : Array.Empty<string>();
+        });
+    }
+
+    public async Task<bool> ShowDeleteSelectedContentWindowAsync(FileOrFolderItem item)
+    {
+        var itemType = item.IsFile ? "file" : "folder";
+        var messageBoxResult = await presentationService.ShowMessageBoxAsync(
+            "Delete from all backups",
+            $"Are you sure you want to delete this {itemType} from all backups?",
+            new List<IUICommand> { new UICommand("Yes"), new UICommand("No") });
+        return messageBoxResult == ContentDialogResult.Primary;
+    }
+
+    public async Task<string?> ShowRenameFavoriteWindowAsync(BrowserFavoriteItem favorite)
+    {
+        return await App.MainWindow.DispatcherQueue.EnqueueAsync(async () =>
+        {
+            var nameBox = new TextBox
+            {
+                Text = favorite.Name,
+                Header = "Name",
+                MinWidth = 320
+            };
+
+            var dialog = new ContentDialog
+            {
+                XamlRoot = App.MainWindow.Content.XamlRoot,
+                Title = "Rename favorite",
+                Content = nameBox,
+                PrimaryButtonText = "Rename",
+                CloseButtonText = "Cancel",
+                DefaultButton = ContentDialogButton.Primary
+            };
+
+            var result = await dialog.ShowAsync();
+            return result == ContentDialogResult.Primary ? nameBox.Text : null;
+        });
+    }
+
+    public async Task ShowFileDetailsAsync(FileDetails fileDetails)
+    {
+        await App.MainWindow.DispatcherQueue.EnqueueAsync(async () =>
+        {
+            var dialog = new ContentDialog
+            {
+                XamlRoot = App.MainWindow.Content.XamlRoot,
+                Title = fileDetails.Name,
+                CloseButtonText = "Close",
+                Content = BuildFileDetailsContent(fileDetails)
+            };
+
+            await dialog.ShowAsync();
+        });
+    }
+
+    private static Grid BuildFileDetailsContent(FileDetails fileDetails)
+    {
+        var grid = new Grid
+        {
+            RowSpacing = 8,
+            ColumnSpacing = 16,
+            MaxWidth = 560
+        };
+        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
+        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });
+
+        AddDetailsRow(grid, "Restore path", fileDetails.RestorePath);
+        AddDetailsRow(grid, "Type", fileDetails.Type);
+        AddDetailsRow(grid, "Size", fileDetails.Size.ToString("N0") + " bytes");
+        AddDetailsRow(grid, "Created", fileDetails.Created.ToString("g"));
+        AddDetailsRow(grid, "Modified", fileDetails.Modified.ToString("g"));
+        AddDetailsRow(grid, "Available versions", string.Join(Environment.NewLine, fileDetails.AvailableVersions.Select(x => $"{x.Id} - {x.CreationDate:g}")));
+
+        return grid;
+    }
+
+    private static void AddDetailsRow(Grid grid, string label, string value)
+    {
+        var row = grid.RowDefinitions.Count;
+        grid.RowDefinitions.Add(new RowDefinition { Height = GridLength.Auto });
+
+        var labelBlock = new TextBlock
+        {
+            Text = label,
+            FontWeight = Microsoft.UI.Text.FontWeights.SemiBold
+        };
+        Grid.SetRow(labelBlock, row);
+        Grid.SetColumn(labelBlock, 0);
+
+        var valueBlock = new TextBlock
+        {
+            Text = value,
+            TextWrapping = Microsoft.UI.Xaml.TextWrapping.Wrap
+        };
+        Grid.SetRow(valueBlock, row);
+        Grid.SetColumn(valueBlock, 1);
+
+        grid.Children.Add(labelBlock);
+        grid.Children.Add(valueBlock);
+    }
+}
